Lab 4 - Quote Generator 1.0.0.
Simple project to create a quote generator. We were provided with the quotes in a const array and asked to create the rest.

Much of the work is simple. Currently I am troubleshooting why my javascript is not able to use the function I've made to call inspirations.length. For some reason the result of this is undefined.

My guess is that it has something to do with it being a const, holding multiple string parts (quote, author, date) rather than being a regular array.